+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftBlockCreateUseCase.java
@@ -1,6 +1,6 @@
 package nettee.blolet.article.application.usecase;
 
-import nettee.draft.draftblock.domain.DraftBlock;
+import nettee.blolet.article.domain.DraftBlock;
 
 public interface DraftBlockCreateUseCase {
         DraftBlock create(String userId, DraftBlock draftBlock);
