+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftBlockReadByStatusUseCase.java
@@ -1,7 +1,7 @@
 package nettee.blolet.article.application.usecase;
 
-import nettee.blolet.article.domain.type.DraftBlockStatus;
-import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockSummary;
+import nettee.blolet.article.domain.sub.DraftBlockStatus;
+import nettee.blolet.article.readmodel.DraftBlockReadModels.DraftBlockSummary;
 
 import java.util.List;
