+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftBlockUpdateUseCase.java
@@ -1,6 +1,6 @@
 package nettee.blolet.article.application.usecase;
 
-import nettee.draft.draftblock.domain.DraftBlock;
+import nettee.blolet.article.domain.DraftBlock;
 
 public interface DraftBlockUpdateUseCase {
     DraftBlock update(String userId, DraftBlock draftBlock);
