+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftImageCreateUseCase.java
@@ -1,6 +1,6 @@
 package nettee.blolet.article.application.usecase;
 
-import nettee.draft.domain.DraftImage;
+import nettee.blolet.article.domain.DraftImage;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface DraftImageCreateUseCase {
