+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftReadByStatusesUseCase.java
@@ -1,7 +1,7 @@
 package nettee.blolet.article.application.usecase;
 
-import nettee.draft.readmodel.DraftReadModels.DraftSummary;
-import nettee.blolet.article.domain.type.DraftStatus;
+import nettee.blolet.article.domain.sub.DraftStatus;
+import nettee.blolet.article.readmodel.DraftReadModels.DraftSummary;
 
 import java.util.List;
 import java.util.Set;
