+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftUpdateUseCase.java
@@ -1,6 +1,6 @@
 package nettee.blolet.article.application.usecase;
 
-import nettee.draft.domain.Draft;
+import nettee.blolet.article.domain.Draft;
 
 public interface DraftUpdateUseCase {
     Draft updateDraft(String userId, Draft draft);
